modified:   constant/common.ts 	modified:   helper/account.helper.ts 	modified:   helper/addCart.helper.ts 	modified:   helper/card.helper.ts 	new file:   helper/rating.helper.ts 	modified:   helper/response.helper.ts     	new file:   migrations/20240528094057-ratingTable.js 	modified:   models/accountModel.ts 	modified:   models/addCartModel.ts 	modified:   models/index.ts 	new file:   models/ratingModel.ts 	modified:   src/controller/transationControler.ts 	modified:   src/controller/userController.ts 	"modified:   src/routes/userRoutes.ts"

+++ models/addCartModel.ts
@@ -1,6 +1,5 @@
-import { Sequelize, Model, DataTypes } from "sequelize";
+import { DataTypes, Model, Sequelize } from "sequelize";
 import { TABLES } from "../constant/common";
-import ProductModel from "./productModel";
 
 class AddCartModel extends Model {
   quantity: any;
@@ -8,13 +7,13 @@ class AddCartModel extends Model {
     this.init(
       {
         id: {
-          type: DataTypes.BIGINT,
+          type: DataTypes.INTEGER,
           allowNull: false,
           autoIncrement: true,
           primaryKey: true,
         },
         userId: {
-          type: DataTypes.BIGINT,
+          type: DataTypes.INTEGER,
           allowNull: false,
         },
